+++ src/main/java/plugily/projects/buildbattle/events/ChatEvents.java
@@ -86,8 +86,10 @@ public class ChatEvents implements Listener {
         .replace("%pts%", Integer.toString(gameArena.getCurrentTheme().getDifficulty().getPointsReward())));
     gameArena.getPlayersPoints().put(event.getPlayer(), gameArena.getPlayersPoints().getOrDefault(event.getPlayer(), 0)
         + gameArena.getCurrentTheme().getDifficulty().getPointsReward());
-    plugin.getUserManager().getUser(gameArena.getCurrentBuilder())
-        .addStat(StatsStorage.StatisticType.LOCAL_GUESS_THE_BUILD_POINTS, gameArena.getCurrentTheme().getDifficulty().getPointsReward());
+    if(gameArena.getWhoGuessed().isEmpty()) {
+      gameArena.getPlayersPoints().put(gameArena.getCurrentBuilder(), gameArena.getPlayersPoints().getOrDefault(event.getPlayer(), 0)
+          + gameArena.getCurrentTheme().getDifficulty().getPointsReward());
+    }
     org.bukkit.Bukkit.getScheduler().runTaskLater(plugin, () -> {
       gameArena.addWhoGuessed(event.getPlayer());
       gameArena.recalculateLeaderboard();